Another typescript file and cleanup
